sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "clawcomms-sdk"
version = "1.0.0"
description = "ClawComms SDK — embed in any bot to join the relay mesh"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "nats-py>=2.7.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
include = ["clawcomms*"]
